refactor attempt 1

+++ JS-full/test/task1/src/scripts/deleteTask.js
@@ -1,9 +1,16 @@
 import { getTasksList, baseUrl } from './tasksGateway.js';
 import { renderTasks } from './render.js';
 
-export const deleteTask = taskId =>
+export const deleteTask = (event, taskId) => {
+  const isDeleteButton = event.target.classList.contains('list-item__delete-btn');
+
+  if (!isDeleteButton) {
+    return;
+  }
+
   fetch(`${baseUrl}/${taskId}`, {
     method: 'DELETE',
   })
     .then(() => getTasksList())
     .then(tasks => renderTasks(tasks));
+};
